Group2 Authorization 추가 완료

+++ src/main/java/com/precapstone/fiveguys_backend/api/group2/Group2Service.java
@@ -1,5 +1,6 @@
 package com.precapstone.fiveguys_backend.api.group2;
 
+import com.precapstone.fiveguys_backend.api.auth.JwtTokenProvider;
 import com.precapstone.fiveguys_backend.api.folder2.Folder2Service;
 import com.precapstone.fiveguys_backend.entity.Group2;
 import com.precapstone.fiveguys_backend.exception.ControlledException;
@@ -7,15 +8,22 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import static com.precapstone.fiveguys_backend.exception.errorcode.Group2ErrorCode.*;
+import static com.precapstone.fiveguys_backend.exception.errorcode.UserErrorCode.USER_AUTHORIZATION_FAILED;
 
 @Service
 @RequiredArgsConstructor
 public class Group2Service {
     private final Group2Repository group2Repository;
     private final Folder2Service folder2Service;
+    private final JwtTokenProvider jwtTokenProvider;
 
-    public Group2 create(Group2CreateDTO group2CreateDTO) {
+    public Group2 create(Group2CreateDTO group2CreateDTO, String accessToken) {
         var folder2 = folder2Service.readFolder2(group2CreateDTO.getFolder2Id());
+
+        // [보안] 데이터의 주인이 호출한 API인지 accessToken을 통해 확인
+        var userId = jwtTokenProvider.getUserIdFromToken(accessToken);
+        if(!folder2.getUser().getUserId().equals(userId))
+            throw new ControlledException(USER_AUTHORIZATION_FAILED);
 
         // 1. [예외처리] 본인 Folder2 안에 Group2가 이미 존재하는 경우
         var groups = folder2.getGroup2s();
@@ -32,15 +40,26 @@ public class Group2Service {
         return group2;
     }
 
-    public Group2 readGroup2(Long group2Id) {
+    public Group2 readGroup2(Long group2Id, String accessToken) {
         var group2 = group2Repository.findById(group2Id)
                 .orElseThrow(()-> new ControlledException(GROUP2_NOT_FOUND));
+
+        // [보안] 데이터의 주인이 호출한 API인지 accessToken을 통해 확인
+        var userId = jwtTokenProvider.getUserIdFromToken(accessToken);
+        if(!group2.getFolder2().getUser().getUserId().equals(userId))
+            throw new ControlledException(USER_AUTHORIZATION_FAILED);
 
         return group2;
     }
 
-    public Group2 update(Group2UpdateDTO group2UpdateDTO) {
-        var group2 = readGroup2(group2UpdateDTO.getGroup2Id());
+    public Group2 update(Group2UpdateDTO group2UpdateDTO, String accessToken) {
+        var group2 = readGroup2(group2UpdateDTO.getGroup2Id(), accessToken);
+
+        // [보안] 데이터의 주인이 호출한 API인지 accessToken을 통해 확인
+        // ※ 이미 readGroup2()에서 인증을 거치지만 형식상 추가
+        var userId = jwtTokenProvider.getUserIdFromToken(accessToken);
+        if(!group2.getFolder2().getUser().getUserId().equals(userId))
+            throw new ControlledException(USER_AUTHORIZATION_FAILED);
 
         if(group2UpdateDTO.getName() != null)
             group2.setName(group2UpdateDTO.getName());
@@ -59,8 +78,14 @@ public class Group2Service {
         return group2;
     }
 
-    public Group2 delete(Long group2Id) {
-        var group2 = readGroup2(group2Id);
+    public Group2 delete(Long group2Id, String accessToken) {
+        var group2 = readGroup2(group2Id, accessToken);
+
+        // [보안] 데이터의 주인이 호출한 API인지 accessToken을 통해 확인
+        // ※ 이미 readGroup2()에서 인증을 거치지만 형식상 추가
+        var userId = jwtTokenProvider.getUserIdFromToken(accessToken);
+        if(!group2.getFolder2().getUser().getUserId().equals(userId))
+            throw new ControlledException(USER_AUTHORIZATION_FAILED);
 
         group2Repository.deleteById(group2Id);
         return group2;
